+++ tests/std/test.lst
@@ -156,6 +156,7 @@ tests\Dev11_1150223_shared_mutex
 tests\Dev11_1158803_regex_thread_safety
 tests\Dev11_1180290_filesystem_error_code
 tests\GH_000140_adl_proof_comparison
+tests\GH_000140_adl_proof_views
 tests\GH_000177_forbidden_aliasing
 tests\GH_000178_uniform_int
 tests\GH_000342_filebuf_close
